merge

+++ src/console/logstorage/dlt-logstorage-common.c
@@ -3,7 +3,7 @@
  * This code is developed by Advanced Driver Information Technology.
  * Copyright of Advanced Driver Information Technology, Bosch and DENSO.
  *
- * This file is part of GENIVI Project Dlt - Diagnostic Log and Trace console apps.
+ * This file is part of COVESA Project Dlt - Diagnostic Log and Trace console apps.
  *
  *
  * \copyright
@@ -17,7 +17,7 @@
  * \author Frederic Berat <fberat@de.adit-jv.com> ADIT 2015
  *
  * \file dlt-logstorage-common.c
- * For further information see http://www.genivi.org/.
+ * For further information see http://www.covesa.org/.
  */
 
 /*******************************************************************************
@@ -104,7 +104,7 @@ int get_default_event_type(void)
 
 void set_default_event_type(long type)
 {
-    g_options.event_type = type;
+    g_options.event_type = (int) type;
 }
 
 char *get_default_path(void)
@@ -285,7 +285,7 @@ static DltControlMsgBody *prepare_message_body(DltControlMsgBody **body,
     serv = (DltServiceOfflineLogstorage *)(*body)->data;
 
     serv->service_id = DLT_SERVICE_ID_OFFLINE_LOGSTORAGE;
-    serv->connection_type = conn_type;
+    serv->connection_type = (uint8_t) conn_type;
     /* mount_point is DLT_MOUNT_PATH_MAX + 1 long,
      * and the memory is already zeroed.
      */
